Move test for issue 5579 under tests/ui/crashes
